Cyclotron_test/toy-insights/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
redis
pytest
pytest-cov
python-dotenv